playlistbackend/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
Deprecated==1.2.13
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
packaging==21.3
pyparsing==3.0.9
pytz==2022.4
redis==4.3.4
requests==2.28.1
six==1.16.0
spotipy==2.20.0
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.12
whitenoise==6.2.0
wrapt==1.14.1
ytmusicapi==0.23.0
